fix(shell): route the startup fullscreen fallback through the F11 transition tracker
On macOS the startup fallback started an asynchronous fullscreen transition
without registering its target, so an F11 pressed during the animation read
isFullScreen() === false and requested fullscreen again instead of leaving
it. The pending-target tracker now lives in its own service
(native-fullscreen-transitions.ts) and both the F11 toggle and the startup
fallback go through it.
