+++ apps/electron-backend/src/app/events/window.events.ts
@@ -13,6 +13,7 @@ import {
     WINDOW_TOGGLE_FULLSCREEN,
     WINDOW_TOGGLE_MAXIMIZE,
 } from '@iptvnator/shared/interfaces';
+import { toggleFullScreen } from '../services/native-fullscreen-transitions';
 
 interface WindowState {
     isMaximized: boolean;
@@ -68,87 +69,13 @@ ipcMain.handle(WINDOW_TOGGLE_MAXIMIZE, (event): WindowState => {
     };
 });
 
-/**
- * Native fullscreen transition still in flight, per window.
- *
- * setFullScreen() completes asynchronously (the macOS animation, Linux
- * window managers) and isFullScreen() keeps reporting the old value
- * meanwhile, so two quick F11 presses would both read the same stale state
- * and request the same target — an enter-then-exit would land fullscreen.
- * The next toggle is therefore decided against the pending target.
- *
- * The entry can never govern F11 for good: it is dropped when a transition
- * event reports the target state, a transition that lands on the OTHER
- * state re-issues the target (the queued request may have been dropped by
- * the platform), and an entry older than the transition timeout is ignored
- * outright, since no native transition takes that long — a request that
- * produced no event at all is then simply forgotten.
- */
-interface PendingFullScreenTransition {
-    target: boolean;
-    requestedAt: number;
-}
-
-/** Longest a native transition is trusted to still be in flight. */
-export const FULLSCREEN_TRANSITION_TIMEOUT_MS = 2000;
-
-const pendingFullScreenTransitions = new WeakMap<
-    Electron.BrowserWindow,
-    PendingFullScreenTransition
->();
-const windowsWithTransitionListeners = new WeakSet<Electron.BrowserWindow>();
-
-function getPendingFullScreenTarget(
-    win: Electron.BrowserWindow,
-    now: number
-): boolean | undefined {
-    const pending = pendingFullScreenTransitions.get(win);
-
-    if (!pending) {
-        return undefined;
-    }
-
-    if (now - pending.requestedAt > FULLSCREEN_TRANSITION_TIMEOUT_MS) {
-        pendingFullScreenTransitions.delete(win);
-        return undefined;
-    }
-
-    return pending.target;
-}
-
-function requestFullScreen(
-    win: Electron.BrowserWindow,
-    target: boolean,
-    now: number
-): void {
-    if (!windowsWithTransitionListeners.has(win)) {
-        windowsWithTransitionListeners.add(win);
-        const settle = () => {
-            const pending = pendingFullScreenTransitions.get(win);
-            if (!pending || win.isDestroyed()) {
-                return;
-            }
-
-            pendingFullScreenTransitions.delete(win);
-            if (win.isFullScreen() !== pending.target) {
-                // An earlier transition landed, not the requested one. Ask
-                // again rather than trusting the platform to have queued
-                // it: a second identical request is a no-op at worst.
-                requestFullScreen(win, pending.target, Date.now());
-            }
-        };
-        win.on('enter-full-screen', settle);
-        win.on('leave-full-screen', settle);
-    }
-
-    pendingFullScreenTransitions.set(win, { target, requestedAt: now });
-    win.setFullScreen(target);
-}
-
 // F11. This is the exit path from a fullscreen launch on Windows/Linux,
 // where the title bar is hidden and the custom controls hide themselves
-// while fullscreen. Like the maximize toggle it reports the requested state
-// and leaves WINDOW:STATE_CHANGED authoritative.
+// while fullscreen. The transition tracker decides against the pending
+// target while a transition (including the startup fullscreen) is still in
+// flight, so two quick presses stay an enter-then-exit. Like the maximize
+// toggle it reports the requested state and leaves WINDOW:STATE_CHANGED
+// authoritative.
 ipcMain.handle(WINDOW_TOGGLE_FULLSCREEN, (event): WindowState => {
     const win = getSenderWindow(event);
 
@@ -156,15 +83,9 @@ ipcMain.handle(WINDOW_TOGGLE_FULLSCREEN, (event): WindowState => {
         return getWindowState(win);
     }
 
-    const now = Date.now();
-    const shouldFullScreen = !(
-        getPendingFullScreenTarget(win, now) ?? win.isFullScreen()
-    );
-    requestFullScreen(win, shouldFullScreen, now);
-
     return {
         isMaximized: win.isMaximized(),
-        isFullScreen: shouldFullScreen,
+        isFullScreen: toggleFullScreen(win),
     };
 });
